feat(deleteProduct): add status 500

+++ backend/controllers/product.controller.js
@@ -20,6 +20,13 @@ export const getProducts = async(req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  if(!mongoose.Types.ObjectId.isValid(id)){
+    return res.status(404).json({
+      success: false,
+      message: "Invalid Product Id"
+    })
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({
@@ -28,9 +35,9 @@ export const deleteProduct = async (req, res) => {
     })
   } catch (error) {
     console.log("Error in deleting product:", error.message);
-    res.status(404).json({
+    res.status(500).json({
       success: false,
-      message: "Product not found"
+      message: "Server Error"
     })
   }
 }
